Added improved command to install programs

+++ Modules/Start-Installation.psm1
@@ -1,22 +1,31 @@
 function Start-Installation
 {
     param(
+        # The name of this program
+        [Parameter(Mandatory=$true)]
+        [String]
+        $Name,
         # The Uri to download the executable
         [Parameter(Mandatory=$true)]
         [String]
         $Uri,
-        # The desired filename with the extension
-        [Parameter(Mandatory=$true)]
-        [String]
-        $Filename,
         # Arguments to pass onto the executable
         [Parameter(Mandatory=$true)]
         [String[]]
-        $Arguments
+        $Arguments,
+        # The registry key that makes the application auto start
+        [String]
+        $StartKey,
+        # The registry value that makes the application auto start
+        [String]
+        $StartValue,
+        # If the temporary file should be removed after installation
+        [switch]
+        $DeleteFile
     )
 
     # Create the path that we are going to use
-    $file = "$env:TEMP\$Filename"
+    $file = "$env:TEMP\$Name.Installer.exe"
     # Remove the existing temp file if it exists
     Remove-Item -Path $file -Force -ErrorAction SilentlyContinue
     # Download the executable file and save it to %TEMP%
@@ -24,5 +33,5 @@ function Start-Installation
     # Start the process with a clean Path and wait for it to finish
     Start-Process -FilePath $file -ArgumentList $Arguments -UseNewEnvironment -Wait
     # Finally, remove the temp file that we used
-    Remove-Item -Path $file -Force
+    if (-Not $DeleteFile) { Remove-Item -Path $file -Force }
 }
